allow script context to be lazily initialized for newly created scripts
