+++ src/OmniSharp.Script/ScriptProjectProvider.cs
@@ -14,7 +14,7 @@ using OmniSharp.Roslyn.Utilities;
 
 namespace OmniSharp.Script
 {
-    public class ScriptHelper
+    public class ScriptProjectProvider
     {
         private const string BinderFlagsType = "Microsoft.CodeAnalysis.CSharp.BinderFlags";
         private const string TopLevelBinderFlagsProperty = "TopLevelBinderFlags";
@@ -49,7 +49,7 @@ namespace OmniSharp.Script
         private readonly ILogger _logger;
         private readonly bool _isDesktopClr;
 
-        public ScriptHelper(ScriptOptions scriptOptions, IOmniSharpEnvironment env, ILoggerFactory loggerFactory, bool isDesktopClr)
+        public ScriptProjectProvider(ScriptOptions scriptOptions, IOmniSharpEnvironment env, ILoggerFactory loggerFactory, bool isDesktopClr)
         {
             _scriptOptions = scriptOptions ?? throw new ArgumentNullException(nameof(scriptOptions));
             _env = env ?? throw new ArgumentNullException(nameof(env));
